+++ src/app/(app)/movies/[id]/components/MovieTrailerButton.tsx
@@ -8,9 +8,11 @@ export const MovieTrailerButton = () => {
 
     if (modalElement) modalElement.showModal();
   };
-
   return (
-    <button className="btn" onClick={handleClick}>
+    <button
+      className="flex w-full items-center justify-center rounded-md border border-transparent bg-indigo-400 px-8 py-3 text-base font-medium text-white hover:bg-indigo-500 focus:outline-none focus:ring-2 focus:ring-indigo-300 focus:ring-offset-2 focus:ring-offset-gray-50"
+      onClick={handleClick}
+    >
       Watch Trailer
     </button>
   );
