+++ src/app/(app)/movies/[id]/components/WatchLaterButton.tsx
@@ -4,7 +4,6 @@ import { useTransition } from "react";
 import { addMovieToWatchList, removeMovieFromWatchList } from "../actions";
 import { useAuthUser } from "~/hooks/useAuthUser";
 import { useRouter } from "next/navigation";
-import { revalidatePath } from "next/cache";
 
 type Props = {
   movieId: number;
@@ -30,7 +29,11 @@ export const WatchLaterButton = ({ movieId, isInWatchList }: Props) => {
   };
 
   return (
-    <button onClick={handleClick} className="relative" disabled={isPending}>
+    <button
+      onClick={handleClick}
+      className="relative flex w-full items-center justify-center rounded-md border border-transparent bg-indigo-600 px-8 py-3 text-base font-medium text-white hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 focus:ring-offset-gray-50"
+      disabled={isPending}
+    >
       {isInWatchList ? "Remove from" : "Add to"} Watch Later
       {isPending && <Loader />}
     </button>
